core/pyproject.toml
[project]
name = "autoflow-ai"
version = "0.0.1.dev33"
description = "Framework to index and search your vector data, make your data ready for AI apps, developed by TiDB."
authors = [
    { name = "Mini256", email = "minianter@foxmail.com" },
    { name = "sykp241095", email = "sykp241095@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.10"

# Dependencies

dependencies = [
    "litellm>=1.61.11",
    "llama-index-core>=0.12.23.post2",
    "llama-index-llms-litellm>=0.3.0",
    "llama-index-readers-file>=0.4.6",
    "deepdiff>=8.2.0",
    "pytidb==0.0.4.dev1",
    "markdownify>=0.13.1",
    "playwright>=1.20.0",
    "dspy>=2.6.6",
    "tokenizers>=0.21.0",
    "mypy>=1.15.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-dotenv>=0.5.2",
    "pre-commit>=4.1.0",
    "pip>=25.0.1",
    "notebook>=7.3.2",
    "markdownify>=0.13.1",
    "playwright>=1.20.0",
    "dspy>=2.6.6",
    "tokenizers>=0.21.0",
    "llama-index-readers-file>=0.4.6",
    "mypy>=1.15.0",
    "numpy>=1.26.4",
    "ipywidgets>=8.1.6",
]

# Check

[tool.pyright]
include = ["autoflow"]

# Test

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
env_files = ["./tests/.env"]

# Build

[tool.setuptools]
packages = ["autoflow"]

[build-system]
requires = ["hatchling==1.26.3"]
build-backend = "hatchling.build"

[tool.hatch.build]
exclude = [
    "/.*",
    "/dist",
    "/docs",
    "/tests",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["autoflow"]
